Removed caching surveys, should be a seperate pull request
