+++ response_operations_ui/controllers/survey_controllers.py
@@ -3,7 +3,7 @@ import logging
 import requests
 from structlog import wrap_logger
 
-from response_operations_ui import app, get_surveys_dict
+from response_operations_ui import app
 from response_operations_ui.exceptions.exceptions import ApiError
 
 logger = wrap_logger(logging.getLogger(__name__))
@@ -32,13 +32,22 @@ def get_survey(short_name):
     return response.json()
 
 
+def get_surveys_dict():
+    logger.debug('Retrieving surveys list')
+    url = f'{app.config["BACKSTAGE_API_URL"]}/v1/survey/surveys'
+    response = requests.get(url)
+    if response.status_code != 200:
+        raise ApiError(response)
+    logger.debug('Successfully retrieved surveys list')
+    try:
+        return {survey['id']: survey for survey in response.json()}
+    except ValueError:
+        logger.exception("Failed to decode survey list")
+
+
 def get_survey_short_name_by_id(survey_id):
     try:
-        return app.surveys_dict[survey_id]['shortName']
-    except KeyError:
-        app.surveys_dict = get_surveys_dict() or app.surveys_dict
-
-    try:
-        app.surveys_dict[survey_id]['shortName']
+        return get_surveys_dict()[survey_id]['shortName']
     except KeyError:
         logger.exception("failed to resolve survey short name", survey_id=survey_id)
+        return 'Unavailable'
