+++ tests/test_data/message/messages.json
@@ -45,9 +45,63 @@
     ],
     "msg_from": "Jordon Dutch",
     "msg_id": "7de3749b-7bb4-4085-8593-4150fff9ff58",
-    "msg_to": [
-     "BRES"
+    "msg_to": ["GROUP"],
+    "ru_id": "dfd3890e-33cd-4e83-bce0-295ddfc91924",
+    "sent_date": "2018-02-07 09:26:11.303877",
+    "subject": "Q3 Statistics",
+    "survey": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87",
+    "thread_id": "7de3749b-7bb4-4085-8593-4150fff9ff58",
+   "level": "info",
+   "service": "response-operations-ui",
+   "created_at": "2018-02-07T15:311518017501"
+  },
+  {
+    "@msg_to": [
+     {
+      "emailAddress": "",
+      "firstName": "BRES",
+      "id": "BRES",
+      "lastName": "",
+      "sampleUnitType": "BI",
+      "status": "",
+      "telephone": ""
+     }
     ],
+    "@ru_id": {
+     "associations": [
+      {
+       "enrolments": [
+        {
+         "enrolmentStatus": "ENABLED",
+         "name": "Business Register and Employment Survey",
+         "surveyId": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87"
+        }
+       ],
+       "partyId": "1b16c90b-15db-4fda-ad3d-fab265f787d8"
+      }
+     ],
+     "id": "dfd3890e-33cd-4e83-bce0-295ddfc91924",
+     "name": "Acme Studios LTD",
+     "sampleSummaryId": null,
+     "sampleUnitRef": "49900000497",
+     "sampleUnitType": "B"
+    },
+    "_links": {
+     "self": {
+      "href": "http://localhost:5050/message/7de3749b-7bb4-4085-8593-4150fff9ff58"
+     }
+    },
+    "body": "testing44444",
+    "collection_case": "6d84ce94-a432-44fa-9ccf-47da5a1171aa",
+    "collection_exercise": "",
+    "labels": [
+     "INBOX",
+     "SENT",
+     "UNREAD"
+    ],
+    "msg_from": "Jordon Dutch",
+    "msg_id": "7de3749b-7bb4-4085-8593-4150fff9ff58",
+    "msg_to": [""],
     "ru_id": "dfd3890e-33cd-4e83-bce0-295ddfc91924",
     "sent_date": "2018-02-07 09:26:11.303877",
     "subject": "Q3 Statistics",
@@ -57,4 +111,5 @@
    "service": "response-operations-ui",
    "created_at": "2018-02-07T15:311518017501"
   }
-]}
+]
+}
