+++ tests/test_data/message/threads.json
@@ -1,75 +1,110 @@
 {
-    "_links": {
-        "first": {
-            "href": "http://localhost:5050/messages"
-        },
-        "next": {
-            "href": "http://localhost:5050/messages?page=2&limit=20"
-        },
-        "self": {
-            "href": "http://localhost:5050/messages?page=1&limit=20"
-        }
+  "_links": {
+    "first": {
+      "href": "http://localhost:5050/messages"
     },
-    "messages": [
+    "next": {
+      "href": "http://localhost:5050/messages?page=2&limit=20"
+    },
+    "self": {
+      "href": "http://localhost:5050/messages?page=1&limit=20"
+    }
+  },
+  "messages": [
+    {
+      "@msg_from": {
+        "emailAddress": "johnexample@example.com",
+        "firstName": "John",
+        "id": "ff4537df-2097-4a73-a530-e98dba7bf28f",
+        "lastName": "Example",
+        "sampleUnitType": "BI",
+        "status": "",
+        "telephone": ""
+      },
+      "@msg_to": [
         {
-            "@msg_from": {
-                "emailAddress": "",
-                "firstName": "John",
-                "id": "ff4537df-2097-4a73-a530-e98dba7bf28f",
-                "lastName": "Soth",
-                "sampleUnitType": "BI",
-                "status": "",
-                "telephone": ""
-            },
-            "@msg_to": [
-                {
-                    "associations": [
-                        {
-                            "enrolments": [
-                                {
-                                    "enrolmentStatus": "ENABLED",
-                                    "name": "Survey Name",
-                                    "surveyId": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87"
-                                }
-                            ],
-                            "partyId": "1f5e1d68-2a4c-4698-8086-e23c0b98923f",
-                            "sampleUnitRef": "50012345678"
-                        }
-                    ],
-                    "emailAddress": "name@email.co.uk",
-                    "firstName": "FirstName",
-                    "id": "ef7737df-2097-4a73-a530-e98dba7bf28f",
-                    "lastName": "LastName",
-                    "sampleUnitType": "BI",
-                    "status": "ACTIVE",
-                    "telephone": "07832323234"
-                }
-            ],
-            "@ru_id":{
-                "id": "f1a5e99c-8edf-489a-9c72-6cabe6c387fc",
-                "name": "Apple",
-                "sampleUnitRef": "50012345678"
-                },
-            "_links": {
-                "self": {
-                    "href": "http://localhost:5050/message/ae46748b-c6e6-4859-a57a-86e01db2dcbc"
-                }
-            },
-            "body": "Test uuid",
-            "collection_case": "ACollectionCase",
-            "collection_exercise": "",
-            "labels": [
-                "INBOX"
-            ],
-            "sent_date": "2017-10-03 15:51:32.961321",
-            "msg_from": "BRES",
-            "msg_id": "ae46748b-c6e6-4859-a57a-86e01db2dcbc",
-            "msg_to": [
-                "ef7737df-2097-4a73-a530-e98dba7bf28f"
-            ],
-            "ru_id": "f1a5e99c-8edf-489a-9c72-6cabe6c387fc",
-            "subject": "This is the message subject",
-            "survey": "f235e99c-8edf-489a-9c72-6cabe6c387fc",
-            "thread_id": "ae46748b-c6e6-4859-a57a-86e01db2dcbc"
+          "emailAddress": "",
+          "firstName": "",
+          "id": "ef7737df-2097-4a73-a530-e98dba7bf28f",
+          "lastName": ""
         }
-]}
+      ],
+      "@ru_id": {
+        "id": "f1a5e99c-8edf-489a-9c72-6cabe6c387fc",
+        "name": "Apple",
+        "sampleUnitRef": "50012345678"
+      },
+      "_links": {
+        "self": {
+          "href": "http://localhost:5050/message/ae46748b-c6e6-4859-a57a-86e01db2dcbc"
+        }
+      },
+      "body": "Example message body",
+      "collection_case": "ACollectionCase",
+      "collection_exercise": "",
+      "labels": ["INBOX"],
+      "sent_date": "2017-10-03 15:51:32.961321",
+      "msg_from": "ff4537df-2097-4a73-a530-e98dba7bf28f",
+      "msg_id": "ae46748b-c6e6-4859-a57a-86e01db2dcbc",
+      "msg_to": ["GROUP"],
+      "ru_id": "f1a5e99c-8edf-489a-9c72-6cabe6c387fc",
+      "subject": "Message from respondent",
+      "survey": "f235e99c-8edf-489a-9c72-6cabe6c387fc",
+      "thread_id": "ae46748b-c6e6-4859-a57a-86e01db2dcbc"
+    },
+    {
+      "@msg_from": {
+        "emailAddress": "",
+        "firstName": "ONS",
+        "id": "ff4537df-2097-4a73-a530-e98dba7bf28f",
+        "lastName": "Employee"
+      },
+      "@msg_to": [
+        {
+          "associations": [
+            {
+              "enrolments": [
+                {
+                  "enrolmentStatus": "ENABLED",
+                  "name": "Survey Name",
+                  "surveyId": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87"
+                }
+              ],
+              "partyId": "1f5e1d68-2a4c-4698-8086-e23c0b98923f",
+              "sampleUnitRef": "50012345678"
+            }
+          ],
+          "emailAddress": "name@email.co.uk",
+          "firstName": "FirstName",
+          "id": "ef7737df-2097-4a73-a530-e98dba7bf28f",
+          "lastName": "LastName",
+          "sampleUnitType": "BI",
+          "status": "ACTIVE",
+          "telephone": "07832323234"
+        }
+      ],
+      "@ru_id": {
+        "id": "f1a5e99c-8edf-489a-9c72-6cabe6c387fc",
+        "name": "Apple",
+        "sampleUnitRef": "50012345678"
+      },
+      "_links": {
+        "self": {
+          "href": "http://localhost:5050/message/ae46748b-c6e6-4859-a57a-86e01db2dcbc"
+        }
+      },
+      "body": "message body",
+      "collection_case": "",
+      "collection_exercise": "",
+      "labels": ["SENT"],
+      "sent_date": "2017-10-03 15:51:32.961321",
+      "msg_from": "GROUP",
+      "msg_id": "9ecfad50-2ff5-4bea-a997-d73c4faa73ae",
+      "msg_to": ["ff4537df-2097-4a73-a530-e98dba7bf28f"],
+      "ru_id": "f1a5e99c-8edf-489a-9c72-6cabe6c387fc",
+      "subject": "Message from ONS",
+      "survey": "f235e99c-8edf-489a-9c72-6cabe6c387fc",
+      "thread_id": "9ecfad50-2ff5-4bea-a997-d73c4faa73ae"
+    }
+  ]
+}
