+++ tests/views/test_message.py
@@ -8,7 +8,9 @@ from response_operations_ui.controllers.message_controllers import _get_url, sen
 from response_operations_ui.exceptions.exceptions import InternalError
 
 url_sign_in_data = f'{app.config["BACKSTAGE_API_URL"]}/v2/sign-in/'
-get_threads_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/threads'
+url_get_threads_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/threads'
+url_get_surveys_list = f'{app.config["BACKSTAGE_API_URL"]}/v1/survey/surveys'
+
 with open('tests/test_data/message/threads.json') as json_data:
     thread_list = json.load(json_data)
 
@@ -27,20 +29,32 @@ class TestMessage(unittest.TestCase):
         # sign-in to setup the user in the session
         self.app.post("/sign-in", follow_redirects=True, data={"username": "user", "password": "pass"})
 
-    # Test showing that the messages list loads into the website and displays User, business name and subject
+    surveys_list_json = [
+        {
+            "id": "f235e99c-8edf-489a-9c72-6cabe6c387fc",
+            "shortName": "QBS",
+            "longName": "Quarterly Business Survey"
+        }
+    ]
+    # Test showing that the threads list loads into the website and displays User, business name and subject
+
     @requests_mock.mock()
     def test_Message_list(self, mock_request):
-        mock_request.get(get_threads_list, json=thread_list)
+        mock_request.get(url_get_threads_list, json=thread_list)
+        mock_request.get(url_get_surveys_list, json=self.surveys_list_json)
 
         response = self.app.get("/messages")
         self.assertEqual(response.status_code, 200)
         self.assertIn("Apple".encode(), response.data)
-        self.assertIn("John Soth".encode(), response.data)
-        self.assertIn("This is the message subject".encode(), response.data)
+        self.assertIn("50012345678".encode(), response.data)
+        self.assertIn("John Example".encode(), response.data)
+        self.assertIn("QBS Team".encode(), response.data)
+        self.assertIn("Message from respondent".encode(), response.data)
+        self.assertIn("Message from ONS".encode(), response.data)
 
     @requests_mock.mock()
     def test_message_list_fail(self, mock_request):
-        mock_request.get(get_threads_list, status_code=500)
+        mock_request.get(url_get_threads_list, status_code=500)
 
         response = self.app.get("/messages", follow_redirects=True)
 
@@ -49,7 +63,7 @@ class TestMessage(unittest.TestCase):
 
     @requests_mock.mock()
     def test_message_list_empty(self, mock_request):
-        mock_request.get(get_threads_list, json={"messages": []})
+        mock_request.get(url_get_threads_list, json={"messages": []})
 
         response = self.app.get("/messages")
 
@@ -82,7 +96,7 @@ class TestMessage(unittest.TestCase):
     # but instead log the problem and display an empty inbox to the user.
     @requests_mock.mock()
     def test_request_response_malformed(self, mock_request):
-        url = get_threads_list
+        url = url_get_threads_list
         mock_request.get(url, json={})
         response = self.app.get("/messages")
 
@@ -128,7 +142,7 @@ class TestMessage(unittest.TestCase):
     @requests_mock.mock()
     def test_form_submit_with_valid_data(self, mock_request):
         mock_request.post(f'{app.config["BACKSTAGE_API_URL"]}/v1/secure-message/send-message', status_code=201)
-        mock_request.get(get_threads_list, json={}, status_code=200)
+        mock_request.get(url_get_threads_list, json={}, status_code=200)
 
         with app.app_context():
             response = self.app.post("/messages/create-message", data=self.message_form, follow_redirects=True)
